requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.10
langchain-openai>=0.1.17
openai>=1.35.0
Pillow>=10.3.0
numpy>=1.26.0
scikit-image>=0.22.0
scikit-learn>=1.4.0
pydantic>=2.7.1
typing-extensions>=4.12.0


